Create auth middleware and verify endpoints functionality

+++ server/api/v1/db/controllers/ordersControllerDB.js
@@ -9,19 +9,21 @@ export default class OrderControllerDB {
    * @param {number || undefined} id
    * @returns {object} orders array if id is undefined or order object otherwise
    */
-  async read(id = '') {
+  async read(req) {
+    // console.log('REQUEST: ', req)
     let result = {};
-    if (id === '') {
-      const getAllQuery = 'SELECT * FROM orders';
-      const { rows, rowCount } = await this.db.query(getAllQuery);
+    if (!req.params.orderId) {
+      const getAllQuery = 'SELECT * FROM orders WHERE owner_id = $1';
+      const { rows, rowCount } = await this.db.query(getAllQuery, [req.user.userId]);
       result = { rows, rowCount };
+      console.log('resut is: ', result);
     } else {
-      const getQuery = 'SELECT * FROM orders WHERE orderid = $1';
-      const { rows } = await this.db.query(getQuery, [id]);
+      const getQuery = 'SELECT * FROM orders WHERE orderid = $1 AND owner_id = $2';
+      const { rows } = await this.db.query(getQuery, [req.params.orderId, req.user.userId]);
       [result] = rows;
     }
     if (!result) {
-      throw new Error(`order with id ${id} not found`);
+      throw new Error(`order with id ${req.params.orderId} not found`);
     }
     return { message: 'order retrieved successfully', order: result };
   }
@@ -31,20 +33,26 @@ export default class OrderControllerDB {
    * @param {object} order
    * @returns {object} order object
    */
-  async create(order) {
-    const query = `INSERT INTO orders(name, quantity, price, userAddr, created_at, modified_at)
-      VALUES($1, $2, $3, $4, $5, $6)
+  async create(req) {
+    console.log('req:', req.params, req.user);
+    const query = `INSERT INTO orders(name, quantity, price, owner_id, userAddr, created_at, modified_at)
+      VALUES($1, $2, $3, $4, $5, $6, $7)
       returning *`;
     const values = [
-      order.name,
-      order.quantity,
-      order.price,
-      order.userAddr,
+      req.body.name,
+      req.body.quantity,
+      req.body.price,
+      req.user.userId,
+      req.body.userAddr,
       new Date(),
       new Date(),
     ];
-    const { rows } = await this.db.query(query, values);
-    return rows[0];
+    try {
+      const { rows } = await this.db.query(query, values);
+      return { message: 'order created successfully', order: rows[0] };
+    } catch (error) {
+      return error;
+    }
   }
 
   /**
@@ -53,26 +61,27 @@ export default class OrderControllerDB {
    * @param {object} attrs
    * @returns {object} updated order
    */
-  async update(id, attrs) {
-    const findOneQuery = 'SELECT * FROM orders WHERE orderId=$1';
-    const { rows } = await this.db.query(findOneQuery, [id]);
+  async update(req) {
+    const findOneQuery = 'SELECT * FROM orders WHERE orderId=$1 AND owner_id = $2';
+    const { rows } = await this.db.query(findOneQuery, [req.params.orderId, req.user.userId]);
     if (rows.length === 0) {
-      throw new Error(`Order with id ${id} not found`);
+      throw new Error(`Order with id ${req.params.orderId} not found`);
     }
     const updateOneQuery = `UPDATE orders
       SET name=$1, quantity=$2, price=$3, userAddr=$4, modified_at= $5
-      WHERE orderid=$6 returning *`;
+      WHERE orderid=$6 AND owner_id = $7 returning *`;
 
     const values = [
-      attrs.name || rows[0].name,
-      attrs.quantity || rows[0].quantity,
-      attrs.price || rows[0].price,
-      attrs.userAddr || rows[0].userAddr,
+      req.body.name || rows[0].name,
+      req.body.quantity || rows[0].quantity,
+      req.body.price || rows[0].price,
+      req.body.userAddr || rows[0].userAddr,
       new Date(),
-      id,
+      req.params.orderId,
+      req.user.userId,
     ];
     const response = await this.db.query(updateOneQuery, values);
-    return response.rows[0];
+    return { message: 'order updated successfully', order: response.rows[0] };
   }
 
   /**
@@ -80,16 +89,16 @@ export default class OrderControllerDB {
    * @param {object} id
    * @returns {void} return status code 204
    */
-  async delete(id) {
-    if (!id) {
+  async delete(req) {
+    if (!req.params.orderId) {
       throw new Error('orderid not specified');
     }
-    const deleteQuery = 'DELETE FROM orders WHERE orderId=$1 returning *';
-    const { rows } = await this.db.query(deleteQuery, [id]);
+    const deleteQuery = 'DELETE FROM orders WHERE orderId=$1 AND owner_id = $2 returning *';
+    const { rows } = await this.db.query(deleteQuery, [req.params.orderId, req.user.userId]);
     if (!rows[0]) {
-      throw new Error(`order with id ${id} not found`);
+      throw new Error(`order with id ${req.params.orderId} not found`);
     }
-    return { message: `order with id ${id} deleted successfully` };
+    return { message: `order with id ${req.params.orderId} deleted successfully` };
   }
   // async delete(req, res) {
   //   const deleteQuery = 'DELETE FROM orders WHERE orderId=$1 returning *';
